requirements.txt
py-cord==2.5.0
requests==2.31.0
pytz==2023.3.post1
aiohttp==3.8.6
python-dotenv==1.0.0
better_profanity==0.7.0
bs4==0.0.1
gitpython==3.1.42
easy-pil==0.3.0
chat-exporter==2.6.1